Refactor composefs-native related code
Move all composefs-native related stuff into `lib/src/bootc_composefs`.
This might help us to spot diff while merging into main, also should
hopefully create less conflicts
