+++ crates/lib/src/bootc_composefs/mod.rs
@@ -1,1 +1,7 @@
+pub(crate) mod boot;
+pub(crate) mod repo;
+pub(crate) mod rollback;
 pub(crate) mod state;
+pub(crate) mod status;
+pub(crate) mod switch;
+pub(crate) mod update;
