+++ crates/lib/src/cli.rs
@@ -28,16 +28,15 @@ use ostree_ext::sysroot::SysrootLock;
 use schemars::schema_for;
 use serde::{Deserialize, Serialize};
 
-use crate::deploy::{composefs_rollback, RequiredHostSpec};
-use crate::install::{
-    pull_composefs_repo, setup_composefs_bls_boot, setup_composefs_uki_boot, write_composefs_state,
-    BootSetupType, BootType,
-};
+use crate::bootc_composefs::rollback::composefs_rollback;
+use crate::bootc_composefs::switch::switch_composefs;
+use crate::bootc_composefs::update::upgrade_composefs;
+use crate::deploy::RequiredHostSpec;
 use crate::lints;
 use crate::progress_jsonl::{ProgressWriter, RawProgressFd};
 use crate::spec::Host;
 use crate::spec::ImageReference;
-use crate::status::{composefs_booted, composefs_deployment_status};
+use crate::status::composefs_booted;
 use crate::utils::sigpolicy_from_opt;
 
 /// Shared progress options
@@ -908,57 +907,6 @@ fn prepare_for_write() -> Result<()> {
     Ok(())
 }
 
-#[context("Upgrading composefs")]
-async fn upgrade_composefs(_opts: UpgradeOpts) -> Result<()> {
-    // TODO: IMPORTANT Have all the checks here that `bootc upgrade` has for an ostree booted system
-
-    let host = composefs_deployment_status()
-        .await
-        .context("Getting composefs deployment status")?;
-
-    // TODO: IMPORTANT We need to check if any deployment is staged and get the image from that
-    let imgref = host
-        .spec
-        .image
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("No image source specified"))?;
-
-    let (repo, entries, id, fs) = pull_composefs_repo(&imgref.transport, &imgref.image).await?;
-
-    let Some(entry) = entries.into_iter().next() else {
-        anyhow::bail!("No boot entries!");
-    };
-
-    let boot_type = BootType::from(&entry);
-    let mut boot_digest = None;
-
-    match boot_type {
-        BootType::Bls => {
-            boot_digest = Some(setup_composefs_bls_boot(
-                BootSetupType::Upgrade((&fs, &host)),
-                repo,
-                &id,
-                entry,
-            )?)
-        }
-
-        BootType::Uki => {
-            setup_composefs_uki_boot(BootSetupType::Upgrade((&fs, &host)), repo, &id, entry)?
-        }
-    };
-
-    write_composefs_state(
-        &Utf8PathBuf::from("/sysroot"),
-        id,
-        imgref,
-        true,
-        boot_type,
-        boot_digest,
-    )?;
-
-    Ok(())
-}
-
 /// Implementation of the `bootc upgrade` CLI command.
 #[context("Upgrading")]
 async fn upgrade(opts: UpgradeOpts) -> Result<()> {
@@ -1072,7 +1020,7 @@ async fn upgrade(opts: UpgradeOpts) -> Result<()> {
     Ok(())
 }
 
-fn imgref_for_switch(opts: &SwitchOpts) -> Result<ImageReference> {
+pub(crate) fn imgref_for_switch(opts: &SwitchOpts) -> Result<ImageReference> {
     let transport = ostree_container::Transport::try_from(opts.transport.as_str())?;
     let imgref = ostree_container::ImageReference {
         transport,
@@ -1083,66 +1031,6 @@ fn imgref_for_switch(opts: &SwitchOpts) -> Result<ImageReference> {
     let target = ImageReference::from(target);
 
     return Ok(target);
-}
-
-#[context("Composefs Switching")]
-async fn switch_composefs(opts: SwitchOpts) -> Result<()> {
-    let target = imgref_for_switch(&opts)?;
-    // TODO: Handle in-place
-
-    let host = composefs_deployment_status()
-        .await
-        .context("Getting composefs deployment status")?;
-
-    let new_spec = {
-        let mut new_spec = host.spec.clone();
-        new_spec.image = Some(target.clone());
-        new_spec
-    };
-
-    if new_spec == host.spec {
-        println!("Image specification is unchanged.");
-        return Ok(());
-    }
-
-    let Some(target_imgref) = new_spec.image else {
-        anyhow::bail!("Target image is undefined")
-    };
-
-    let (repo, entries, id, fs) =
-        pull_composefs_repo(&"docker".into(), &target_imgref.image).await?;
-
-    let Some(entry) = entries.into_iter().next() else {
-        anyhow::bail!("No boot entries!");
-    };
-
-    let boot_type = BootType::from(&entry);
-    let mut boot_digest = None;
-
-    match boot_type {
-        BootType::Bls => {
-            boot_digest = Some(setup_composefs_bls_boot(
-                BootSetupType::Upgrade((&fs, &host)),
-                repo,
-                &id,
-                entry,
-            )?)
-        }
-        BootType::Uki => {
-            setup_composefs_uki_boot(BootSetupType::Upgrade((&fs, &host)), repo, &id, entry)?
-        }
-    };
-
-    write_composefs_state(
-        &Utf8PathBuf::from("/sysroot"),
-        id,
-        &target_imgref,
-        true,
-        boot_type,
-        boot_digest,
-    )?;
-
-    Ok(())
 }
 
 /// Implementation of the `bootc switch` CLI command.
