+++ crates/lib/src/spec.rs
@@ -11,7 +11,7 @@ use ostree_ext::{container::OstreeImageReference, oci_spec};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::install::BootType;
+use crate::bootc_composefs::boot::BootType;
 use crate::{k8sapitypes, status::Slot};
 
 const API_VERSION: &str = "org.containers.bootc/v1";
