+++ crates/lib/src/status.rs
@@ -3,43 +3,27 @@ use std::collections::VecDeque;
 use std::io::IsTerminal;
 use std::io::Read;
 use std::io::Write;
-use std::str::FromStr;
 use std::sync::OnceLock;
 
 use anyhow::{Context, Result};
 use bootc_kernel_cmdline::Cmdline;
-use bootc_utils::try_deserialize_timestamp;
 use canon_json::CanonJsonSerialize;
-use cap_std_ext::cap_std;
-use cap_std_ext::cap_std::ambient_authority;
-use cap_std_ext::cap_std::fs::Dir;
 use fn_error_context::context;
 use ostree::glib;
 use ostree_container::OstreeImageReference;
 use ostree_ext::container as ostree_container;
-use ostree_ext::container::deploy::ORIGIN_CONTAINER;
 use ostree_ext::container_utils::ostree_booted;
-use ostree_ext::containers_image_proxy;
 use ostree_ext::keyfileext::KeyFileExt;
 use ostree_ext::oci_spec;
 use ostree_ext::oci_spec::image::Digest;
 use ostree_ext::oci_spec::image::ImageConfiguration;
 use ostree_ext::sysroot::SysrootLock;
 
-use ostree_ext::oci_spec::image::ImageManifest;
 use ostree_ext::ostree;
-use tokio::io::AsyncReadExt;
 
+use crate::bootc_composefs::status::composefs_deployment_status;
 use crate::cli::OutputFormat;
-use crate::composefs_consts::{
-    COMPOSEFS_CMDLINE, COMPOSEFS_STAGED_DEPLOYMENT_FNAME, COMPOSEFS_TRANSIENT_STATE_DIR,
-    ORIGIN_KEY_BOOT, ORIGIN_KEY_BOOT_TYPE, STATE_DIR_RELATIVE,
-};
-use crate::deploy::get_sorted_bls_boot_entries;
-use crate::deploy::get_sorted_uki_boot_entries;
-use crate::install::BootType;
-use crate::install::EFIVARFS;
-use crate::spec::Bootloader;
+use crate::composefs_consts::COMPOSEFS_CMDLINE;
 use crate::spec::ImageStatus;
 use crate::spec::{BootEntry, BootOrder, Host, HostSpec, HostStatus, HostType};
 use crate::spec::{ImageReference, ImageSignature};
@@ -398,249 +382,6 @@ pub(crate) fn get_status(
         ty,
     };
     Ok((deployments, host))
-}
-
-/// imgref = transport:image_name
-#[context("Getting container info")]
-async fn get_container_manifest_and_config(
-    imgref: &String,
-) -> Result<(ImageManifest, oci_spec::image::ImageConfiguration)> {
-    let config = containers_image_proxy::ImageProxyConfig::default();
-    let proxy = containers_image_proxy::ImageProxy::new_with_config(config).await?;
-
-    let img = proxy.open_image(&imgref).await.context("Opening image")?;
-
-    let (_, manifest) = proxy.fetch_manifest(&img).await?;
-    let (mut reader, driver) = proxy.get_descriptor(&img, manifest.config()).await?;
-
-    let mut buf = Vec::with_capacity(manifest.config().size() as usize);
-    buf.resize(manifest.config().size() as usize, 0);
-    reader.read_exact(&mut buf).await?;
-    driver.await?;
-
-    let config: oci_spec::image::ImageConfiguration = serde_json::from_slice(&buf)?;
-
-    Ok((manifest, config))
-}
-
-#[context("Getting bootloader")]
-fn get_bootloader() -> Result<Bootloader> {
-    let efivarfs = match Dir::open_ambient_dir(EFIVARFS, ambient_authority()) {
-        Ok(dir) => dir,
-        // Most likely using BIOS
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(Bootloader::Grub),
-        Err(e) => Err(e).context(format!("Opening {EFIVARFS}"))?,
-    };
-
-    const EFI_LOADER_INFO: &str = "LoaderInfo-4a67b082-0a4c-41cf-b6c7-440b29bb8c4f";
-
-    match efivarfs.read_to_string(EFI_LOADER_INFO) {
-        Ok(loader) => {
-            if loader.to_lowercase().contains("systemd-boot") {
-                return Ok(Bootloader::Systemd);
-            }
-
-            return Ok(Bootloader::Grub);
-        }
-
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(Bootloader::Grub),
-
-        Err(e) => Err(e).context(format!("Opening {EFI_LOADER_INFO}"))?,
-    }
-}
-
-#[context("Getting composefs deployment metadata")]
-async fn boot_entry_from_composefs_deployment(
-    origin: tini::Ini,
-    verity: String,
-) -> Result<BootEntry> {
-    let image = match origin.get::<String>("origin", ORIGIN_CONTAINER) {
-        Some(img_name_from_config) => {
-            let ostree_img_ref = OstreeImageReference::from_str(&img_name_from_config)?;
-            let imgref = ostree_img_ref.imgref.to_string();
-            let img_ref = ImageReference::from(ostree_img_ref);
-
-            // The image might've been removed, so don't error if we can't get the image manifest
-            let (image_digest, version, architecture, created_at) =
-                match get_container_manifest_and_config(&imgref).await {
-                    Ok((manifest, config)) => {
-                        let digest = manifest.config().digest().to_string();
-                        let arch = config.architecture().to_string();
-                        let created = config.created().clone();
-                        let version = manifest
-                            .annotations()
-                            .as_ref()
-                            .and_then(|a| a.get(oci_spec::image::ANNOTATION_VERSION).cloned());
-
-                        (digest, version, arch, created)
-                    }
-
-                    Err(e) => {
-                        tracing::debug!("Failed to open image {img_ref}, because {e:?}");
-                        ("".into(), None, "".into(), None)
-                    }
-                };
-
-            let timestamp = created_at.and_then(|x| try_deserialize_timestamp(&x));
-
-            let image_status = ImageStatus {
-                image: img_ref,
-                version,
-                timestamp,
-                image_digest,
-                architecture,
-            };
-
-            Some(image_status)
-        }
-
-        // Wasn't booted using a container image. Do nothing
-        None => None,
-    };
-
-    let boot_type = match origin.get::<String>(ORIGIN_KEY_BOOT, ORIGIN_KEY_BOOT_TYPE) {
-        Some(s) => BootType::try_from(s.as_str())?,
-        None => anyhow::bail!("{ORIGIN_KEY_BOOT} not found"),
-    };
-
-    let e = BootEntry {
-        image,
-        cached_update: None,
-        incompatible: false,
-        pinned: false,
-        store: None,
-        ostree: None,
-        composefs: Some(crate::spec::BootEntryComposefs {
-            verity,
-            boot_type,
-            bootloader: get_bootloader()?,
-        }),
-        soft_reboot_capable: false,
-    };
-
-    return Ok(e);
-}
-
-#[context("Getting composefs deployment status")]
-pub(crate) async fn composefs_deployment_status() -> Result<Host> {
-    let composefs_state = composefs_booted()?
-        .ok_or_else(|| anyhow::anyhow!("Failed to find composefs parameter in kernel cmdline"))?;
-    let composefs_digest = &composefs_state.digest;
-
-    let sysroot = cap_std::fs::Dir::open_ambient_dir("/sysroot", cap_std::ambient_authority())
-        .context("Opening sysroot")?;
-    let deployments = sysroot
-        .read_dir(STATE_DIR_RELATIVE)
-        .with_context(|| format!("Reading sysroot {STATE_DIR_RELATIVE}"))?;
-
-    let host_spec = HostSpec {
-        image: None,
-        boot_order: BootOrder::Default,
-    };
-
-    let mut host = Host::new(host_spec);
-
-    let staged_deployment_id = match std::fs::File::open(format!(
-        "{COMPOSEFS_TRANSIENT_STATE_DIR}/{COMPOSEFS_STAGED_DEPLOYMENT_FNAME}"
-    )) {
-        Ok(mut f) => {
-            let mut s = String::new();
-            f.read_to_string(&mut s)?;
-
-            Ok(Some(s))
-        }
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(None),
-        Err(e) => Err(e),
-    }?;
-
-    // NOTE: This cannot work if we support both BLS and UKI at the same time
-    let mut boot_type: Option<BootType> = None;
-
-    for depl in deployments {
-        let depl = depl?;
-
-        let depl_file_name = depl.file_name();
-        let depl_file_name = depl_file_name.to_string_lossy();
-
-        // read the origin file
-        let config = depl
-            .open_dir()
-            .with_context(|| format!("Failed to open {depl_file_name}"))?
-            .read_to_string(format!("{depl_file_name}.origin"))
-            .with_context(|| format!("Reading file {depl_file_name}.origin"))?;
-
-        let ini = tini::Ini::from_string(&config)
-            .with_context(|| format!("Failed to parse file {depl_file_name}.origin as ini"))?;
-
-        let boot_entry =
-            boot_entry_from_composefs_deployment(ini, depl_file_name.to_string()).await?;
-
-        // SAFETY: boot_entry.composefs will always be present
-        let boot_type_from_origin = boot_entry.composefs.as_ref().unwrap().boot_type;
-
-        match boot_type {
-            Some(current_type) => {
-                if current_type != boot_type_from_origin {
-                    anyhow::bail!("Conflicting boot types")
-                }
-            }
-
-            None => {
-                boot_type = Some(boot_type_from_origin);
-            }
-        };
-
-        if depl.file_name() == composefs_digest.as_ref() {
-            host.spec.image = boot_entry.image.as_ref().map(|x| x.image.clone());
-            host.status.booted = Some(boot_entry);
-            continue;
-        }
-
-        if let Some(staged_deployment_id) = &staged_deployment_id {
-            if depl_file_name == staged_deployment_id.trim() {
-                host.status.staged = Some(boot_entry);
-                continue;
-            }
-        }
-
-        host.status.rollback = Some(boot_entry);
-    }
-
-    // Shouldn't really happen, but for sanity nonetheless
-    let Some(boot_type) = boot_type else {
-        anyhow::bail!("Could not determine boot type");
-    };
-
-    let boot_dir = sysroot.open_dir("boot").context("Opening boot dir")?;
-
-    match boot_type {
-        BootType::Bls => {
-            host.status.rollback_queued = !get_sorted_bls_boot_entries(&boot_dir, false)?
-                .first()
-                .ok_or(anyhow::anyhow!("First boot entry not found"))?
-                .options
-                .as_ref()
-                .ok_or(anyhow::anyhow!("options key not found in bls config"))?
-                .contains(composefs_digest.as_ref());
-        }
-
-        BootType::Uki => {
-            let mut s = String::new();
-
-            host.status.rollback_queued = !get_sorted_uki_boot_entries(&boot_dir, &mut s)?
-                .first()
-                .ok_or(anyhow::anyhow!("First boot entry not found"))?
-                .body
-                .chainloader
-                .contains(composefs_digest.as_ref())
-        }
-    };
-
-    if host.status.rollback_queued {
-        host.spec.boot_order = BootOrder::Rollback
-    };
-
-    Ok(host)
 }
 
 /// Implementation of the `bootc status` CLI command.
